Organised tests; new attack-data tests
